// src/lib/database/types.ts
export interface NewsArticleInsert {
  url: string;
  title: string;
  content?: string | null;
  portal: string;
  publishedAt?: Date | null;
  contentHash: string;
}